Separate app and main
